# crypto-trading-ai/shared/openapi/api.yaml
openapi: 3.1.0
info:
  title: "Crypto Trading AI API"
  version: "0.1.0"
  description: "Generated from backend runtime OpenAPI. Do not edit manually."
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/connections/exchanges:
    get:
      summary: "List Exchanges"
      responses:
        "200":
          description: "Successful Response"
    post:
      summary: "Save Exchange"
      responses:
        "200":
          description: "Successful Response"
        "422":
          description: "Validation Error"
  /api/v1/connections/exchanges/test:
    post:
      summary: "Test Exchange"
      responses:
        "200":
          description: "Successful Response"
        "422":
          description: "Validation Error"
...